ULTRADEFINITIVO VERSION 456 FINAL NO HATE

+++ src/presentation/controller/DeleteMusicController.java
@@ -1,6 +1,8 @@
 package presentation.controller;
 
+import business.Owner;
 import business.SongManager;
+import business.entities.Song;
 import presentation.view.BottomBarPanel;
 import presentation.view.DeleteMusicPanelView;
 
@@ -30,15 +32,19 @@ public class DeleteMusicController implements ActionListener {
     @Override
     public void actionPerformed(ActionEvent e) {
         if (e.getActionCommand().equals(DeleteMusicPanelView.DELETE)) {
-
+            Song song = getSongByName(deleteMusicPanelView.getNameSong());
             if (BottomBarPanel.getSong() != null) {
                 if (!BottomBarPanel.getSong().getTitle().equals(deleteMusicPanelView.getNameSong())) {
                     if (deleteMusicPanelView.getNameSong() != null) {
-
-                        if (SongManager.DeleteSong(deleteMusicPanelView.getNameSong())) {
-                            JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE SUCCESSFULLY", "", JOptionPane.INFORMATION_MESSAGE);
+                        song = getSongByName(deleteMusicPanelView.getNameSong());
+                        if (song.getOwner().equals(Owner.getUser().getName())) {
+                            if (SongManager.DeleteSong(deleteMusicPanelView.getNameSong())) {
+                                JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE SUCCESSFULLY", "", JOptionPane.INFORMATION_MESSAGE);
+                            } else {
+                                JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE UNSUCCESSFULLY", "", JOptionPane.INFORMATION_MESSAGE);
+                            }
                         } else {
-                            JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE UNSUCCESSFULLY", "", JOptionPane.INFORMATION_MESSAGE);
+                            JOptionPane.showMessageDialog(deleteMusicPanelView, "CANNOT DELETE A SONG OF WHICH YOU ARE NOT OWNER", "Following errors were found", JOptionPane.WARNING_MESSAGE);
                         }
 
                     } else {
@@ -48,12 +54,21 @@ public class DeleteMusicController implements ActionListener {
                     JOptionPane.showMessageDialog(deleteMusicPanelView, "SONG IS PLAYING", "Following errors were found", JOptionPane.WARNING_MESSAGE);
                 }
             } else {
-                if (SongManager.DeleteSong(deleteMusicPanelView.getNameSong())) {
-                    JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE SUCCESSFULLY", "", JOptionPane.INFORMATION_MESSAGE);
+                if (song.getOwner().equals(Owner.getUser().getName())) {
+                    if (SongManager.DeleteSong(deleteMusicPanelView.getNameSong())) {
+                        JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE SUCCESSFULLY", "", JOptionPane.INFORMATION_MESSAGE);
+                    } else {
+                        JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE UNSUCCESSFULLY", "", JOptionPane.INFORMATION_MESSAGE);
+                    }
                 } else {
-                    JOptionPane.showMessageDialog(deleteMusicPanelView, "DELETE UNSUCCESSFULLY", "", JOptionPane.INFORMATION_MESSAGE);
+                    JOptionPane.showMessageDialog(deleteMusicPanelView, "CANNOT DELETE A SONG OF WHICH YOU ARE NOT OWNER", "Following errors were found", JOptionPane.WARNING_MESSAGE);
                 }
+
             }
         }
+    }
+
+    private Song getSongByName (String name) {
+        return SongManager.SelectSong(name);
     }
 }
